Pretty Print Update

+++ pysdql/query/tpch/__init__.py
@@ -10,7 +10,7 @@ from pysdql.query.util import sdql_to_df, pandas_to_df, compare_dataframe
 def tpch_query(qindex=1, execution_mode=0, threads_count=1) -> bool:
     done = [1, 6]
 
-    if isinstance(qindex, float):
+    if isinstance(qindex, int):
         if qindex not in done:
             print(f'Query {qindex} has not been verified.')
             return False
